feat(ItemWriteDiaryImage.kt): DiaryImage 클래스에 isSelected 값을 바꾸는 함수 추가
